Fix circular import by moving AsyncConfigEntryAuth to auth.py
__init__.py imported coordinator.py, which imported AsyncConfigEntryAuth
back from __init__.py, causing a circular import on startup. Moved
AsyncConfigEntryAuth to auth.py; coordinator now imports it from there
directly alongside PostNLGraphql from graphql.py.
