+++ custom_components/postnl/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import time
 
 import requests
 import urllib3
@@ -10,10 +9,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady, HomeAssistantError
 from homeassistant.helpers import device_registry as dr, entity_registry as er
 
-from .auth import PostNLAuth, PostNLAuthError
+from .auth import AsyncConfigEntryAuth
 from .const import DOMAIN, PLATFORMS
 from .coordinator import PostNLCoordinator
-from .graphql import PostNLGraphql
 from .login_api import PostNLLoginAPI
 
 _LOGGER = logging.getLogger(__name__)
@@ -82,54 +80,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class AsyncConfigEntryAuth:
-    """Manage PostNL tokens stored in a config entry."""
-
-    def __init__(self, hass: HomeAssistant, entry: ConfigEntry) -> None:
-        self._hass = hass
-        self._entry = entry
-
-    @property
-    def access_token(self) -> str:
-        return self._entry.data["token"]["access_token"]
-
-    async def check_and_refresh_token(self) -> str:
-        token = self._entry.data.get("token")
-
-        if not token or "access_token" not in token:
-            self._entry.async_start_reauth(self._hass)
-            raise HomeAssistantError("No valid token in config entry, reauth required")
-
-        if time.time() < token.get("expires_at", 0) - 30:
-            return token["access_token"]
-
-        _LOGGER.debug("Access token expired, refreshing")
-        refresh_token = token.get("refresh_token")
-        if refresh_token:
-            try:
-                new_token = await PostNLAuth.async_refresh_token(refresh_token)
-                self._hass.config_entries.async_update_entry(
-                    self._entry,
-                    data={**self._entry.data, "token": new_token},
-                )
-                return new_token["access_token"]
-            except PostNLAuthError as err:
-                _LOGGER.debug("Token refresh failed, falling back to re-login: %s", err)
-
-        username = self._entry.data.get("username")
-        password = self._entry.data.get("password")
-        if username and password:
-            try:
-                new_token = await PostNLAuth(username, password).async_login()
-                self._hass.config_entries.async_update_entry(
-                    self._entry,
-                    data={**self._entry.data, "token": new_token},
-                )
-                return new_token["access_token"]
-            except PostNLAuthError as err:
-                _LOGGER.debug("Re-login failed, triggering reauth: %s", err)
-
-        self._entry.async_start_reauth(self._hass)
-        raise HomeAssistantError("Unable to obtain a valid token")
