+++ custom_components/postnl/auth.py
@@ -1,11 +1,19 @@
+from __future__ import annotations
+
 import base64
 import hashlib
+import logging
 import re
 import secrets
 import time
 from urllib.parse import parse_qs, urlparse
 
 import aiohttp
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
+from homeassistant.exceptions import HomeAssistantError
+
+_LOGGER = logging.getLogger(__name__)
 
 _BASE = "https://login.postnl.nl"
 _TENANT = "101112a0-4a0f-4bbb-8176-2f1b2d370d7c"
@@ -254,3 +262,54 @@ class PostNLAuth:
     def _json_value(body: str, key: str) -> str | None:
         m = re.search(rf'"{re.escape(key)}"\s*:\s*"([^"]+)"', body)
         return m.group(1) if m else None
+
+
+class AsyncConfigEntryAuth:
+    """Manage PostNL tokens stored in a config entry."""
+
+    def __init__(self, hass: HomeAssistant, entry: ConfigEntry) -> None:
+        self._hass = hass
+        self._entry = entry
+
+    @property
+    def access_token(self) -> str:
+        return self._entry.data["token"]["access_token"]
+
+    async def check_and_refresh_token(self) -> str:
+        token = self._entry.data.get("token")
+
+        if not token or "access_token" not in token:
+            self._entry.async_start_reauth(self._hass)
+            raise HomeAssistantError("No valid token in config entry, reauth required")
+
+        if time.time() < token.get("expires_at", 0) - 30:
+            return token["access_token"]
+
+        _LOGGER.debug("Access token expired, refreshing")
+        refresh_token = token.get("refresh_token")
+        if refresh_token:
+            try:
+                new_token = await PostNLAuth.async_refresh_token(refresh_token)
+                self._hass.config_entries.async_update_entry(
+                    self._entry,
+                    data={**self._entry.data, "token": new_token},
+                )
+                return new_token["access_token"]
+            except PostNLAuthError as err:
+                _LOGGER.debug("Token refresh failed, falling back to re-login: %s", err)
+
+        username = self._entry.data.get("username")
+        password = self._entry.data.get("password")
+        if username and password:
+            try:
+                new_token = await PostNLAuth(username, password).async_login()
+                self._hass.config_entries.async_update_entry(
+                    self._entry,
+                    data={**self._entry.data, "token": new_token},
+                )
+                return new_token["access_token"]
+            except PostNLAuthError as err:
+                _LOGGER.debug("Re-login failed, triggering reauth: %s", err)
+
+        self._entry.async_start_reauth(self._hass)
+        raise HomeAssistantError("Unable to obtain a valid token")
